requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
Django==4.0.2
django-otp==1.1.3
django-otp-twilio==1.0.1
django-otp-yubikey==1.0.1
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
idna==3.3
phonenumbers==8.12.42
pycryptodome==3.14.1
qrcode==7.3.1
requests==2.27.1
sqlparse==0.4.2
twilio==7.5.1
tzdata==2021.5
urllib3==1.26.8
YubiOTP==1.0.0
